finished testing config object, started adding documentation
